Refactor: split up AttributeV3Mapper
