+++ molgenis-api-metadata/src/main/java/org/molgenis/api/metadata/v3/AttributeRequestMapper.java
@@ -1,26 +1,16 @@
 package org.molgenis.api.metadata.v3;
 
-import static java.util.Collections.emptyList;
 import static java.util.Objects.requireNonNull;
-import static java.util.stream.Collectors.toList;
 import static org.molgenis.api.metadata.v3.MetadataUtils.getStringValue;
 import static org.molgenis.api.metadata.v3.MetadataUtils.setAttributeType;
 import static org.molgenis.api.metadata.v3.MetadataUtils.setBooleanValue;
 import static org.molgenis.api.metadata.v3.MetadataUtils.setEnumOptions;
 import static org.molgenis.api.metadata.v3.MetadataUtils.setSequenceNumber;
-import static org.molgenis.data.meta.AttributeType.ONE_TO_MANY;
-import static org.molgenis.data.meta.AttributeType.getValueString;
-import static org.molgenis.data.meta.model.AttributeMetadata.DESCRIPTION;
 import static org.molgenis.data.meta.model.AttributeMetadata.IS_READ_ONLY;
 import static org.molgenis.data.meta.model.AttributeMetadata.IS_UNIQUE;
-import static org.molgenis.data.meta.model.AttributeMetadata.LABEL;
 import static org.molgenis.util.i18n.LanguageService.getLanguageCodes;
-import static org.springframework.web.servlet.support.ServletUriComponentsBuilder.fromCurrentRequestUri;
 
-import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.Iterables;
 import com.google.common.collect.Maps;
-import java.net.URI;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -34,26 +24,14 @@ import org.molgenis.api.convert.SortConverter;
 import org.molgenis.api.data.SortMapper;
 import org.molgenis.api.metadata.v3.exception.InvalidKeyException;
 import org.molgenis.api.metadata.v3.exception.UnsupportedFieldException;
-import org.molgenis.api.metadata.v3.model.AttributeResponse;
-import org.molgenis.api.metadata.v3.model.AttributeResponseData;
-import org.molgenis.api.metadata.v3.model.AttributeResponseData.Builder;
-import org.molgenis.api.metadata.v3.model.AttributesResponse;
-import org.molgenis.api.metadata.v3.model.Category;
 import org.molgenis.api.metadata.v3.model.CreateAttributeRequest;
 import org.molgenis.api.metadata.v3.model.CreateEntityTypeRequest;
 import org.molgenis.api.metadata.v3.model.I18nValue;
 import org.molgenis.api.metadata.v3.model.Range;
-import org.molgenis.api.model.Sort.Order.Direction;
-import org.molgenis.api.model.response.LinksResponse;
-import org.molgenis.api.model.response.PageResponse;
-import org.molgenis.api.support.LinksUtils;
 import org.molgenis.data.DataConverter;
-import org.molgenis.data.Entity;
 import org.molgenis.data.EntityManager;
 import org.molgenis.data.InvalidValueTypeException;
 import org.molgenis.data.Repository;
-import org.molgenis.data.Sort;
-import org.molgenis.data.Sort.Order;
 import org.molgenis.data.UnknownAttributeException;
 import org.molgenis.data.UnknownEntityTypeException;
 import org.molgenis.data.UnknownRepositoryException;
@@ -64,15 +42,10 @@ import org.molgenis.data.meta.model.AttributeFactory;
 import org.molgenis.data.meta.model.AttributeMetadata;
 import org.molgenis.data.meta.model.EntityType;
 import org.molgenis.data.meta.model.EntityTypeMetadata;
-import org.molgenis.data.util.EntityTypeUtils;
-import org.springframework.context.i18n.LocaleContextHolder;
 import org.springframework.stereotype.Component;
-import org.springframework.web.util.UriComponentsBuilder;
 
 @Component
-class AttributeV3Mapper {
-  private static final String ATTRIBUTES = "attributes";
-
+class AttributeRequestMapper {
   private final AttributeFactory attributeFactory;
   private final MetaDataService metaDataService;
   private final SortMapper sortMapper;
@@ -80,7 +53,7 @@ class AttributeV3Mapper {
   private final EntityManager entityManager;
   private final EntityTypeMetadata entityTypeMetadata;
 
-  AttributeV3Mapper(
+  AttributeRequestMapper(
       AttributeFactory attributeFactory,
       MetaDataService metaDataService,
       SortMapper sortMapper,
@@ -93,119 +66,6 @@ class AttributeV3Mapper {
     this.sortConverter = requireNonNull(sortConverter);
     this.entityManager = requireNonNull(entityManager);
     this.entityTypeMetadata = requireNonNull(entityTypeMetadata);
-  }
-
-  AttributesResponse mapAttributes(Attributes attributes, int size, int number, int total) {
-    return AttributesResponse.create(
-        LinksUtils.createLinksResponse(number, size, total),
-        attributes.getAttributes().stream()
-            .map(attribute -> this.mapAttribute(attribute, false))
-            .collect(toList()),
-        PageResponse.create(size, attributes.getTotal(), attributes.getTotal() / size, number));
-  }
-
-  List<AttributeResponse> mapInternal(
-      Iterable<org.molgenis.data.meta.model.Attribute> allAttributes, boolean i18n) {
-    List<AttributeResponse> result = new ArrayList<>();
-    for (Attribute attr : allAttributes) {
-      result.add(mapAttribute(attr, i18n));
-    }
-    return result;
-  }
-
-  AttributeResponse mapAttribute(Attribute attr, boolean i18n) {
-    AttributeResponseData attribute = mapInternal(attr, i18n);
-    return AttributeResponse.create(
-        LinksResponse.create(null, createAttributeResponseUri(attr), null), attribute);
-  }
-
-  private AttributeResponseData mapInternal(Attribute attr, boolean i18n) {
-    Builder builder = AttributeResponseData.builder();
-    builder.setId(attr.getIdentifier());
-    builder.setName(attr.getName());
-    builder.setSequenceNr(attr.getSequenceNumber());
-    builder.setType(getValueString(attr.getDataType()));
-    builder.setIdAttribute(attr.isIdAttribute());
-    builder.setLabelAttribute(attr.isLabelAttribute());
-    builder.setLookupAttributeIndex(attr.getLookupAttributeIndex());
-    if (EntityTypeUtils.isReferenceType(attr)) {
-      builder.setRefEntityType(
-          LinksResponse.create(null, createEntityTypeResponseUri(attr.getRefEntity()), null));
-    }
-    builder.setCascadeDelete(attr.getCascadeDelete());
-    builder.setMappedBy(attr.getMappedBy() != null ? mapAttribute(attr.getMappedBy(), i18n) : null);
-    if (attr.getDataType() == ONE_TO_MANY && attr.isMappedBy()) {
-      builder.setOrderBy(map(attr));
-    }
-    builder.setLabel(attr.getLabel(LocaleContextHolder.getLocale().getLanguage()));
-    builder.setDescription(attr.getDescription(LocaleContextHolder.getLocale().getLanguage()));
-    if (i18n) {
-      builder.setLabelI18n(getI18nAttrLabel(attr));
-      getI18nAttrDesc(attr).ifPresent(builder::setDescriptionI18n);
-    }
-    builder.setNullable(attr.isNillable());
-    builder.setAuto(attr.isAuto());
-    builder.setVisible(attr.isVisible());
-    builder.setUnique(attr.isUnique());
-    builder.setReadOnly(attr.isReadOnly());
-    builder.setAggregatable(attr.isAggregatable());
-    builder.setExpression(attr.getExpression());
-    if (attr.getDataType() == AttributeType.ENUM) {
-      builder.setEnumOptions(attr.getEnumOptions());
-    }
-    if (attr.getDataType() == AttributeType.CATEGORICAL
-        || attr.getDataType() == AttributeType.CATEGORICAL_MREF) {
-      builder.setCategoricalOptions(getCategories(attr.getRefEntity()));
-    }
-    org.molgenis.data.Range range = attr.getRange();
-    if (range != null) {
-      builder.setRange(Range.create(range.getMin(), range.getMax()));
-    }
-    Attribute parent = attr.getParent();
-    builder.setParentAttributeId(parent != null ? parent.getIdentifier() : null);
-    builder.setNullableExpression(attr.getNullableExpression());
-    builder.setVisibleExpression(attr.getVisibleExpression());
-    builder.setValidationExpression(attr.getValidationExpression());
-    builder.setDefaultValue(attr.getDefaultValue());
-
-    return builder.build();
-  }
-
-  private List<Category> getCategories(EntityType entityType) {
-    Repository<Entity> repository =
-        metaDataService
-            .getRepository(entityType)
-            .orElseThrow(() -> new UnknownRepositoryException(entityType.getId()));
-    Attribute idAttribute = entityType.getIdAttribute();
-    Attribute labelAttribute = entityType.getLabelAttribute();
-    return repository
-        .query()
-        .findAll()
-        .map(
-            entity ->
-                Category.builder()
-                    .setId(entity.get(idAttribute))
-                    .setLabel(entity.get(labelAttribute).toString())
-                    .build())
-        .collect(toList());
-  }
-
-  private List<org.molgenis.api.model.Sort> map(Attribute attr) {
-    List<org.molgenis.api.model.Sort> orders;
-
-    Sort sort = attr.getOrderBy();
-    if (sort == null) {
-      orders = emptyList();
-    } else {
-      orders = new ArrayList<>(Iterables.size(sort));
-      for (Order order : sort) {
-        orders.add(
-            org.molgenis.api.model.Sort.create(
-                order.getAttr(), Direction.valueOf(order.getDirection().name())));
-      }
-    }
-
-    return orders;
   }
 
   public Attribute toAttribute(CreateAttributeRequest attributeRequest, EntityType entityType) {
@@ -576,26 +436,6 @@ class AttributeV3Mapper {
     return attribute;
   }
 
-  private URI createAttributeResponseUri(Attribute attr) {
-    UriComponentsBuilder uriComponentsBuilder =
-        fromCurrentRequestUri()
-            .replacePath(null)
-            .path(MetadataApiController.API_META_PATH)
-            .pathSegment(attr.getEntity().getId())
-            .pathSegment(ATTRIBUTES)
-            .pathSegment(attr.getIdentifier());
-    return uriComponentsBuilder.build().toUri();
-  }
-
-  private URI createEntityTypeResponseUri(EntityType entityType) {
-    UriComponentsBuilder uriComponentsBuilder =
-        fromCurrentRequestUri()
-            .replacePath(null)
-            .path(MetadataApiController.API_META_PATH)
-            .pathSegment(entityType.getId());
-    return uriComponentsBuilder.build().toUri();
-  }
-
   private void processI18nLabel(I18nValue i18nValue, Attribute attribute) {
     if (i18nValue != null) {
       attribute.setLabel(i18nValue.getDefaultValue());
@@ -619,20 +459,5 @@ class AttributeV3Mapper {
                     attribute.setDescription(languageCode, translations.get(languageCode)));
       }
     }
-  }
-
-  private I18nValue getI18nAttrLabel(Attribute attr) {
-    String defaultValue = attr.getLabel();
-    ImmutableMap<String, String> translations = MetadataUtils.getI18n(attr, LABEL);
-    return I18nValue.create(defaultValue, translations);
-  }
-
-  private Optional<I18nValue> getI18nAttrDesc(Attribute attr) {
-    String defaultValue = attr.getDescription();
-    if (defaultValue == null) {
-      return Optional.empty();
-    }
-    ImmutableMap<String, String> translations = MetadataUtils.getI18n(attr, DESCRIPTION);
-    return Optional.of(I18nValue.create(defaultValue, translations));
   }
 }
