+++ molgenis-api-metadata/src/test/java/org/molgenis/api/metadata/v3/AttributeRequestMapperTest.java
@@ -1,11 +1,9 @@
 package org.molgenis.api.metadata.v3;
 
-import static java.util.Arrays.asList;
 import static java.util.Collections.singletonList;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
@@ -14,9 +12,6 @@ import static org.molgenis.data.meta.model.AttributeMetadata.IS_READ_ONLY;
 import static org.molgenis.data.meta.model.AttributeMetadata.IS_UNIQUE;
 
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
-import java.net.URI;
-import java.net.URISyntaxException;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
@@ -28,15 +23,10 @@ import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.molgenis.api.convert.SortConverter;
 import org.molgenis.api.data.SortMapper;
-import org.molgenis.api.metadata.v3.model.AttributeResponse;
-import org.molgenis.api.metadata.v3.model.AttributeResponseData;
-import org.molgenis.api.metadata.v3.model.AttributesResponse;
 import org.molgenis.api.metadata.v3.model.CreateAttributeRequest;
 import org.molgenis.api.metadata.v3.model.CreateEntityTypeRequest;
 import org.molgenis.api.metadata.v3.model.I18nValue;
 import org.molgenis.api.metadata.v3.model.Range;
-import org.molgenis.api.model.response.LinksResponse;
-import org.molgenis.api.model.response.PageResponse;
 import org.molgenis.data.EntityManager;
 import org.molgenis.data.Repository;
 import org.molgenis.data.Sort;
@@ -49,11 +39,8 @@ import org.molgenis.data.meta.model.AttributeMetadata;
 import org.molgenis.data.meta.model.EntityType;
 import org.molgenis.data.meta.model.EntityTypeMetadata;
 import org.molgenis.test.AbstractMockitoTest;
-import org.springframework.mock.web.MockHttpServletRequest;
-import org.springframework.web.context.request.RequestContextHolder;
-import org.springframework.web.context.request.ServletRequestAttributes;
 
-class AttributeV3MapperTest extends AbstractMockitoTest {
+class AttributeRequestMapperTest extends AbstractMockitoTest {
 
   @Mock private AttributeFactory attributeFactory;
   @Mock private MetaDataService metaDataService;
@@ -62,12 +49,12 @@ class AttributeV3MapperTest extends AbstractMockitoTest {
   @Mock private EntityManager entityManager;
   @Mock private EntityTypeMetadata entityTypeMetadata;
 
-  private AttributeV3Mapper attributeV3Mapper;
+  private AttributeRequestMapper attributeRequestMapper;
 
   @BeforeEach
   void setUpBeforeEach() {
-    attributeV3Mapper =
-        new AttributeV3Mapper(
+    attributeRequestMapper =
+        new AttributeRequestMapper(
             attributeFactory,
             metaDataService,
             sortMapper,
@@ -80,132 +67,7 @@ class AttributeV3MapperTest extends AbstractMockitoTest {
   void testAttributeV3Mapper() {
     assertThrows(
         NullPointerException.class,
-        () -> new AttributeV3Mapper(null, null, null, null, null, null));
-  }
-
-  @Test
-  void mapAttributes() throws URISyntaxException {
-    MockHttpServletRequest mockHttpServletRequest = new MockHttpServletRequest();
-    mockHttpServletRequest.setRequestURI("/api/metadata/MyEntityType/attributes");
-    mockHttpServletRequest.setQueryString("page=1");
-    RequestContextHolder.setRequestAttributes(new ServletRequestAttributes(mockHttpServletRequest));
-
-    EntityType entityType = mock(EntityType.class);
-    when(entityType.getId()).thenReturn("MyEntityTypeId");
-
-    Attribute attribute = mock(Attribute.class);
-    when(attribute.getIdentifier()).thenReturn("MyAttributeId");
-    when(attribute.getName()).thenReturn("MyAttributeName");
-    when(attribute.getSequenceNumber()).thenReturn(2);
-    when(attribute.getEntity()).thenReturn(entityType);
-    when(attribute.getDataType()).thenReturn(AttributeType.STRING);
-    when(attribute.getLookupAttributeIndex()).thenReturn(null);
-    when(attribute.getCascadeDelete()).thenReturn(null);
-
-    AttributeResponseData attributeResponseData =
-        AttributeResponseData.builder()
-            .setId("MyAttributeId")
-            .setName("MyAttributeName")
-            .setSequenceNr(2)
-            .setType("string")
-            .setIdAttribute(false)
-            .setLabelAttribute(false)
-            .setNullable(false)
-            .setAuto(false)
-            .setVisible(false)
-            .setUnique(false)
-            .setReadOnly(false)
-            .setAggregatable(false)
-            .build();
-
-    AttributeResponse attributeResponse =
-        AttributeResponse.builder()
-            .setLink(
-                LinksResponse.create(
-                    null,
-                    new URI(
-                        "http://localhost/api/metadata/MyEntityTypeId/attributes/MyAttributeId"),
-                    null))
-            .setData(attributeResponseData)
-            .build();
-
-    AttributesResponse attributesResponse =
-        AttributesResponse.builder()
-            .setLinks(
-                LinksResponse.create(
-                    new URI("http://localhost/api/metadata/MyEntityType/attributes?page=0"),
-                    new URI("http://localhost/api/metadata/MyEntityType/attributes?page=1"),
-                    new URI("http://localhost/api/metadata/MyEntityType/attributes?page=2")))
-            .setItems(ImmutableList.of(attributeResponse))
-            .setPage(PageResponse.create(1, 5, 5, 1))
-            .build();
-
-    assertEquals(
-        attributesResponse,
-        attributeV3Mapper.mapAttributes(
-            Attributes.create(ImmutableList.of(attribute), 5), 1, 1, 5));
-  }
-
-  @Test
-  void mapAttribute() throws URISyntaxException {
-    MockHttpServletRequest mockHttpServletRequest = new MockHttpServletRequest();
-    RequestContextHolder.setRequestAttributes(new ServletRequestAttributes(mockHttpServletRequest));
-
-    EntityType entityType = mock(EntityType.class);
-    when(entityType.getId()).thenReturn("MyEntityTypeId");
-
-    Attribute attribute = mock(Attribute.class);
-    when(attribute.getIdentifier()).thenReturn("MyAttributeId");
-    when(attribute.getName()).thenReturn("MyAttributeName");
-    doReturn("My Attribute").when(attribute).getLabel();
-    doReturn(null).when(attribute).getLabel("en");
-    doReturn("My Attribute description").when(attribute).getDescription();
-    doReturn(null).when(attribute).getDescription("en");
-    when(attribute.getSequenceNumber()).thenReturn(2);
-    when(attribute.getEntity()).thenReturn(entityType);
-    when(attribute.getDataType()).thenReturn(AttributeType.ENUM);
-    when(attribute.getEnumOptions()).thenReturn(asList("option0", "option1"));
-    when(attribute.getLookupAttributeIndex()).thenReturn(null);
-    when(attribute.getCascadeDelete()).thenReturn(null);
-
-    AttributeResponseData attributeResponseData =
-        AttributeResponseData.builder()
-            .setId("MyAttributeId")
-            .setName("MyAttributeName")
-            .setSequenceNr(2)
-            .setType("enum")
-            .setEnumOptions(ImmutableList.of("option0", "option1"))
-            .setIdAttribute(false)
-            .setLabelAttribute(false)
-            .setNullable(false)
-            .setAuto(false)
-            .setVisible(false)
-            .setUnique(false)
-            .setReadOnly(false)
-            .setAggregatable(false)
-            .setLabelI18n(
-                I18nValue.builder()
-                    .setDefaultValue("My Attribute")
-                    .setTranslations(ImmutableMap.of())
-                    .build())
-            .setDescriptionI18n(
-                I18nValue.builder()
-                    .setDefaultValue("My Attribute description")
-                    .setTranslations(ImmutableMap.of())
-                    .build())
-            .build();
-
-    AttributeResponse attributeResponse =
-        AttributeResponse.builder()
-            .setLink(
-                LinksResponse.create(
-                    null,
-                    new URI(
-                        "http://localhost/api/metadata/MyEntityTypeId/attributes/MyAttributeId"),
-                    null))
-            .setData(attributeResponseData)
-            .build();
-    assertEquals(attributeResponse, attributeV3Mapper.mapAttribute(attribute, true));
+        () -> new AttributeRequestMapper(null, null, null, null, null, null));
   }
 
   @Test
@@ -230,7 +92,7 @@ class AttributeV3MapperTest extends AbstractMockitoTest {
     EntityType entityType = mock(EntityType.class);
 
     List<Attribute> attributes =
-        attributeV3Mapper.toAttributes(
+        attributeRequestMapper.toAttributes(
             singletonList(createAttributeRequest), entityTypeRequest, entityType);
 
     assertAll(
@@ -263,7 +125,7 @@ class AttributeV3MapperTest extends AbstractMockitoTest {
     attributeValueMap.put("id", 1);
     attributeValueMap.put("sequenceNr", 1);
     EntityType entityType = mock(EntityType.class);
-    attributeV3Mapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
+    attributeRequestMapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
 
     verify(attribute).setSequenceNumber(1);
   }
@@ -277,7 +139,7 @@ class AttributeV3MapperTest extends AbstractMockitoTest {
     attributeValueMap.put("id", 1);
     attributeValueMap.put("name", "test");
     EntityType entityType = mock(EntityType.class);
-    attributeV3Mapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
+    attributeRequestMapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
 
     verify(attribute).setName("test");
   }
@@ -297,7 +159,7 @@ class AttributeV3MapperTest extends AbstractMockitoTest {
     attributeValueMap.put("aggregatable", true);
     attributeValueMap.put("unique", true);
     EntityType entityType = mock(EntityType.class);
-    attributeV3Mapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
+    attributeRequestMapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
 
     assertAll(
         () -> verify(attribute).set(IS_READ_ONLY, true),
@@ -321,7 +183,7 @@ class AttributeV3MapperTest extends AbstractMockitoTest {
     attributeValueMap.put("id", 1);
     attributeValueMap.put("range", range);
     EntityType entityType = mock(EntityType.class);
-    attributeV3Mapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
+    attributeRequestMapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
 
     verify(attribute).setRange(new org.molgenis.data.Range(1l, 10l));
   }
@@ -341,7 +203,7 @@ class AttributeV3MapperTest extends AbstractMockitoTest {
     attributeValueMap.put("label", i18n);
     attributeValueMap.put("description", i18n);
     EntityType entityType = mock(EntityType.class);
-    attributeV3Mapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
+    attributeRequestMapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
 
     assertAll(
         () -> verify(attribute).setLabel("default"),
@@ -365,7 +227,7 @@ class AttributeV3MapperTest extends AbstractMockitoTest {
     Map<String, Object> attributeValueMap = new HashMap<>();
     attributeValueMap.put("id", "attribute");
     attributeValueMap.put("parent", "test");
-    attributeV3Mapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
+    attributeRequestMapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
 
     verify(attribute).setParent(parentAttr);
   }
@@ -387,7 +249,7 @@ class AttributeV3MapperTest extends AbstractMockitoTest {
     attributeValueMap.put("id", "attribute");
     attributeValueMap.put("mappedByAttribute", "test");
     EntityType entityType = mock(EntityType.class);
-    attributeV3Mapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
+    attributeRequestMapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
 
     verify(attribute).setMappedBy(mappedByAttr);
   }
@@ -405,7 +267,7 @@ class AttributeV3MapperTest extends AbstractMockitoTest {
     attributeValueMap.put("defaultValue", "defaultValue");
     attributeValueMap.put("validationExpression", "validationExpression");
     EntityType entityType = mock(EntityType.class);
-    attributeV3Mapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
+    attributeRequestMapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
 
     assertAll(
         () -> verify(attribute).setExpression("expression"),
@@ -424,7 +286,7 @@ class AttributeV3MapperTest extends AbstractMockitoTest {
     attributeValueMap.put("id", 1);
     attributeValueMap.put("enumOptions", Arrays.asList("1", "2"));
     EntityType entityType = mock(EntityType.class);
-    attributeV3Mapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
+    attributeRequestMapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
 
     verify(attribute).setEnumOptions(Arrays.asList("1", "2"));
   }
@@ -442,7 +304,7 @@ class AttributeV3MapperTest extends AbstractMockitoTest {
     attributeValueMap.put("id", 1);
     attributeValueMap.put("orderBy", "-id");
     EntityType entityType = mock(EntityType.class);
-    attributeV3Mapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
+    attributeRequestMapper.toAttributes(singletonList(attributeValueMap), entityType);
 
     verify(attribute).setOrderBy(molgenisSort);
   }
@@ -456,7 +318,7 @@ class AttributeV3MapperTest extends AbstractMockitoTest {
     attributeValueMap.put("id", 1);
     attributeValueMap.put("type", "onetomany");
     EntityType entityType = mock(EntityType.class);
-    attributeV3Mapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
+    attributeRequestMapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
 
     verify(attribute).setDataType(AttributeType.ONE_TO_MANY);
   }
@@ -474,7 +336,7 @@ class AttributeV3MapperTest extends AbstractMockitoTest {
     attributeValueMap.put("id", "attribute");
     attributeValueMap.put("refEntityType", "ref");
 
-    attributeV3Mapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
+    attributeRequestMapper.toAttributes(Collections.singletonList(attributeValueMap), entityType);
 
     verify(attribute).setRefEntity(entityType);
   }
